修改正确的HSE晶振设置，修复波特率只能跑一半的问题

+++ Hardware/uart_manager/uart_manager.h
@@ -19,7 +19,7 @@
 #define UART_RECV_QUEUE_MAX_LEN 8
 
 #define  USART1_COM                USART1
-#define  USART1_BAUDRATE           (u32)115200
+#define  USART1_BAUDRATE           115200
 #define  USART1_CLK                RCC_APB2PERIPH_USART1
 #define  USART1_APBxClkCmd         RCC_APB2PeriphClockCmd
